split up class-specific materialize logic
